sqldirstate: write a flat dirstate with just parents for backwards comp
Summary:
This is neccessary for scmprompt to work. I added a config knob because we
may want to disable it if we want to do paranoid mode at some point.

Test Plan: ran tests

Reviewers: #mercurial, durham

Differential Revision: https://phabricator.intern.facebook.com/D3356646

+++ sqldirstate/__init__.py
@@ -7,7 +7,7 @@
 
 testedwith = 'internal'
 
-from sqldirstate import makedirstate, DBFILE, toflat, tosql
+from sqldirstate import makedirstate, DBFILE, toflat, tosql, writefakedirstate
 
 from mercurial import commands, error, extensions, cmdutil, localrepo, util
 from mercurial.i18n import _
@@ -81,8 +81,12 @@ def upgrade(ui, repo):
         tosql(repo.dirstate)
         repo.requirements.add('sqldirstate')
         repo._writerequirements()
-        repo.dirstate._opener.unlink('dirstate')
+        if ui.configbool('sqldirstate', 'fakedirstate', True):
+            writefakedirstate(repo.dirstate)
+        else:
+            repo.dirstate._opener.unlink('dirstate')
         del repo.dirstate
+
     finally:
         wlock.release()
